pyproject.toml
[tool.poetry]
name = "42c-cli"
version = "0.1.0"
description = "CLI for 42Crunch"
authors = ["42crunch <dev@42crunch.com>"]
readme = "README.md"
packages = [{include = "xliic_cli"}]

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.9.0"
requests = "^2.30.0"
pyyaml = "^6.0"
sentry-sdk = "^1.24.0"
jsonpointer = "^2.3"
ruamel-yaml = "^0.17.31"
docker = "^6.1.3"

mitmproxy = {version = "^9.0.1", optional = true}
aiofiles = {version = "^23.1.0", optional = true}


[tool.poetry.group.dev.dependencies]
types-requests = "^2.31.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[[tool.poetry.source]]
name = "42Crunch"
url = "https://repo.42crunch.com/pypi/"
priority = "explicit"


# add entrypoint
[tool.poetry.scripts]
42c = "xliic_cli.__main__:main"

# add extra dependencies
[tool.poetry.extras]
proxy = [
    "mitmproxy", # ^9.0.1
    "aiofiles" # ^23.1.0
]

